requirements.txt
aiohttp==3.11.11
aiosignal==1.3.2
APScheduler==3.9.1
attrs==25.1.0
base58==2.1.1
blinker==1.9.0
borsh-construct==0.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
Flask==3.1.0
frozenlist==1.5.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
hexbytes==0.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.0
pytz==2025.1
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
requests-oauthlib==1.3.1
six==1.17.0
sniffio==1.3.1
solana==0.35.1
solders==0.21.0
tabulate==0.9.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
web3==6.20.3
# Remove websockets explicit dependency, let it be installed as a dependency
Werkzeug==3.1.3
yarl==1.18.3
httpx==0.27.0
tweepy
discord.py
requests_oauthlib>=1.3.0
anthropic>=0.15.0
cairosvg==2.7.1
svgwrite
